--- a/vimebu.go
+++ b/vimebu.go
@@ -4,7 +4,6 @@ import (
 	"bytes"
 	"strconv"
 	"strings"
-	"sync"
 )
 
 const (
@@ -18,27 +17,6 @@ const (
 	equalByte        = byte('=')
 	doubleQuotesByte = byte('"')
 )
-
-// bytesBufferPool is a simple pool to create or retrieve a [bytes.Buffer].
-var bytesBufferPool = sync.Pool{
-	New: func() any {
-		return new(bytes.Buffer)
-	},
-}
-
-// getBuffer acquires a [bytes.Buffer] from the pool.
-func getBuffer() *bytes.Buffer {
-	return bytesBufferPool.Get().(*bytes.Buffer)
-}
-
-// putBuffer resets and returns a [bytes.Buffer] to the pool.
-func putBuffer(b *bytes.Buffer) {
-	if b == nil {
-		return
-	}
-	b.Reset()
-	bytesBufferPool.Put(b)
-}
 
 // Builder is used to efficiently build a VictoriaMetrics metric.
 // It's backed by a bytes.Buffer to minimize memory copying.